+++ imagewin/iwin8.h
@@ -117,8 +117,8 @@ public:
 		ib8->Image_buffer8::put_pixel8(pix, x, y);
 	}
 
-	static void get_gamma(float &r, float &g, float &b);
-	static void set_gamma(float r, float g, float b);
+	static void get_gamma(double &r, double &g, double &b);
+	static void set_gamma(double r, double g, double b);
 
 	std::unique_ptr<unsigned char[]> mini_screenshot();
 };
